Add `initialHours` option and update calculation in Hours class
Added a new property `initialHours` to the `HoursOptions` interface and updated the `calculateSlip()` method of the `Hours` class to use this value for initializing total slip. Also, a console log has been added to display the totalSlip for debugging purposes. Lastly, the `parseArgs()` function now accepts an optional `initialHours` argument, which defaults to 0 if not provided. If the environment variable `TOGGL_INITIAL_HOURS` is set, it will be used as the default value. This change allows for starting the calculation from a specific initial time.

Commit message by mistral:latest